+++ frontend/__test__/CreateFunnel.test.tsx
@@ -1,7 +1,7 @@
 import { ThemeProvider } from '@emotion/react';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import SpaceCreateFunnel from '../src/pages/create/SpaceCreateFunnel';
+import SpaceCreateFunnel from '../src/pages/create/funnel/SpaceCreateFunnel';
 import { theme } from '../src/styles/theme';
 
 describe('스페이스 생성 퍼널 테스트', () => {
