+++ frontend/src/pages/create/funnel/SpaceCreateFunnel.styles.ts
