+++ frontend/src/router/router.tsx
@@ -3,7 +3,7 @@ import Layout from '../components/layout/Layout';
 import DownloadCompletePage from '../pages/complete/DownloadCompletePage';
 import SpaceCreatedCompletePage from '../pages/complete/SpaceCreatedCompletePage';
 import UploadCompletePage from '../pages/complete/UploadCompletePage';
-import SpaceCreateFunnel from '../pages/create/SpaceCreateFunnel';
+import SpaceCreateFunnel from '../pages/create/funnel/SpaceCreateFunnel';
 import DemoHome from '../pages/demo/DemoHome';
 import NetworkErrorPage from '../pages/error/NetworkErrorPage';
 import ImageUploadPage from '../pages/guest/imageUploadPage/ImageUploadPage';
